% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_imap_data.R
\name{cache_imap_data}
\alias{cache_imap_data}
\title{Cache data from Maryland iMap}
\usage{
cache_imap_data(
  data = NULL,
  nm = NULL,
  ...,
  location = NULL,
  name = NULL,
  label = NULL,
  fileext = "gpkg",
  filename = NULL,
  path = NULL,
  prefix = NULL,
  postfix = NULL,
  cache = TRUE,
  pkg = "getdata",
  create = TRUE,
  overwrite = FALSE,
  call = caller_env()
)
}
\arguments{
\item{data}{Data to pass to \code{\link[getdata:format_address_data]{getdata::bind_block_col()}}.
\code{\link[=format_parcel_data]{format_parcel_data()}} only.}

\item{nm}{Layer name identifier used to retrieve url from \code{md_imap_index}
based on the snakecase "nm" column; Default: \code{NULL}.}

\item{...}{Additional parameters passed to \code{\link[getdata:get_esri_data]{getdata::get_esri_data()}}}

\item{location}{A \code{sf}, \code{sfc}, or \code{bbox} object (or other object convertible
with \code{\link[sfext:as_sf]{sfext::as_bbox()}}. Required.}

\item{name}{Name to make file name converted to snake case with
\code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}, e.g. "Residential zoning map" becomes
"residential_zoning_map". If the name includes a file extension it is
assumed that the filename has been provided as the name parameter.}

\item{label}{Label to combine with name converted to snake case with
\code{\link[janitor:make_clean_names]{janitor::make_clean_names()}}. The label is designed to identify the area
or other shared characteristics across multiple data files, maps, or plots.
label is ignored if name is NULL or if name includes a file extension.}

\item{fileext}{File type or extension. Optional if filename or path include a
file extension.}

\item{filename}{File name; if filename is \code{NULL} and path does not include a
file extension, name and file extension are both required.}

\item{path}{Path to file or data directory. Optional. If path includes a file
extension and filename and fileext are both \code{NULL}, the filename and
extension included with path will be used instead. If multiple file
extensions are provided to filename, path, or fileext, \code{make_filename()}
will abort.}

\item{prefix}{File name prefix. "date" adds a date prefix, "time" adds a
date/time prefix; defaults to \code{NULL}.}

\item{postfix}{File name postfix; defaults to \code{NULL}.}

\item{cache}{If \code{TRUE}, path is set to the package cache directory using
\code{\link[filenamr:get_data_dir]{get_data_dir()}}; defaults to \code{FALSE}.}

\item{pkg}{Package name passed to appname parameter of
\code{\link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir()}}}

\item{create}{If \code{FALSE} and path does not exist, return path with a warning.
If \code{TRUE} and \code{\link[rlang:is_interactive]{rlang::is_interactive()}} is \code{TRUE}, ask user if directory
should be created. If the session not interactive and create is \code{TRUE}, a
new directory will be created.}

\item{overwrite}{If \code{TRUE}, remove a file with the same name and path}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=cache_imap_data]{cache_imap_data()}} is a wrapper for \code{\link[=get_imap_data]{get_imap_data()}} and
\code{\link[getdata:cache_location_data]{getdata::cache_location_data()}} that save a layer from Maryland iMap or a
subset of a layer to a local file. By default, the data is saved to the
mapmaryland package cache folder.
}
