% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_md_data.R
\name{format_md_sf}
\alias{format_md_sf}
\alias{format_md_crash_data}
\title{Format Maryland simple feature data}
\usage{
format_md_sf(
  data,
  crs = getOption("mapmaryland.crs", default = 3857),
  erase_water = FALSE,
  erase_data = NULL,
  clean_names = TRUE,
  dTolerance = NULL,
  smooth = FALSE,
  sf_col = "geometry",
  ...
)

format_md_crash_data(data, drop_code = TRUE)
}
\arguments{
\item{data}{Data frame with Maryland vehicular crash data; typically from \code{get_md_crash_data}}

\item{crs}{coordinate reference system. Defaults to
getOption("mapmaryland.crs", default = 3857). This option can be set with
\link{set_mapmaryland_options}.}

\item{erase_water}{If \code{TRUE}, erase any geometry intersecting with
\link{md_water}; Default: \code{FALSE}. If TRUE, and erase_data is not \code{NULL}, \link{md_water}
is combined and unioned with erase_data.}

\item{erase_data}{A \code{sf}, \code{sfc}, or \code{bbox} object with geometry that should be erased
from the data, Default: \code{NULL}}

\item{dTolerance}{numeric; tolerance parameter, specified for all or for each
feature geometry. If you run \code{st_simplify}, the input data is
specified with long-lat coordinates and \code{sf_use_s2()} returns
\code{TRUE}, then the value of \code{dTolerance} must be specified in
meters.}

\item{smooth}{If \code{TRUE}, smooth data with \link[smoothr:smooth]{smoothr::smooth} using default
method and parameters, Default: \code{FALSE}.}

\item{sf_col}{Name to use for output \code{sf} column, Default: 'geometry'.}

\item{...}{Additional parameters passed to \code{format_data}}

\item{drop_code}{If TRUE, drop all columns that end with "code"}
}
\value{
A modified version of the input simple feature data.
}
\description{
Convert the coordinate reference system, convert invalid geometry to valid
geometry with \link[sf:valid]{sf::st_make_valid}, and, optionally, erase water, and clean
names with \link[janitor:clean_names]{janitor::clean_names}. \link{format_md_crash_data} works with a data
frame of crash data from \link{get_md_crash_data}
}
\seealso{
\link[getdata:format_sf_data]{getdata::format_sf_data}
}
