% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imap.R
\name{get_imap_data}
\alias{get_imap_data}
\alias{get_cama_data}
\alias{get_parcel_data}
\alias{get_mihp_data}
\alias{get_water_data}
\title{Get data from Maryland iMap}
\usage{
get_imap_data(location, nm = NULL, crs = 3857, ...)

get_cama_data(location, type = "bldg", ...)

get_parcel_data(location, ...)

get_mihp_data(location, ...)

get_water_data(location, type = "streams", ...)
}
\arguments{
\item{location}{A \code{sf}, \code{sfc}, or \code{bbox} object (or other object convertible
with \link[overedge:as_sf]{overedge::as_bbox}. Required.}

\item{crs}{Coordinate reference system to return.}

\item{...}{Additional parameters passed to \link[overedge:get_esri_data]{overedge::get_esri_data}}

\item{type}{See details.}
}
\description{
Get data from Maryland iMap including Computer Assisted Mass Appraisal
(CAMA), parcel data, Maryland Inventory of Historic Properties data, and a
variety of water data.
}
\details{
Options for type parameter:
\itemize{
\item For \link{get_cama_data}, Default: 'bldg'; options include "bldg", "core",
"land", "subarea".
\item For \link{get_water_data}, Default: "streams", options include "streams",
"streams detailed", "federal watersheds", "8 digit watersheds", "12 digit
watersheds", "navigable waterways", "ponds"
}
}
