% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_md_boundary.R
\name{get_md_boundary}
\alias{get_md_boundary}
\title{Get Maryland administrative boundaries}
\usage{
get_md_boundary(name = NULL, type = "county")
}
\arguments{
\item{name}{Name matching a value in the namelsad, namelsad, or geoid columns
of \link{md_counties}, \link{md_census_places}, \link{md_congressional_districts},
\link{md_legislative_districts}, or \link{md_senate_districts},
Default: NULL}

\item{type}{Type, "county", "census place", "congressional district",
"legislative district", or "senate district";
Default: 'county'}
}
\value{
A simple feature object matching the type provided.
}
\description{
Get Maryland counties, U.S. Congressional District, or Census designated
places by name or GeoID.
}
