% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_md_tigris.R
\name{get_md_tigris}
\alias{get_md_tigris}
\title{Use package data or the {tigris} package to get Maryland data from the U.S. Census Bureau}
\usage{
get_md_tigris(
  name = NULL,
  type = "counties",
  crs = 3857,
  erase_water = FALSE,
  ...
)
}
\arguments{
\item{name}{Name matching a value in the namelsad, namelsad, or geoid columns
of \link{md_counties}, \link{md_census_places}, \link{md_congressional_districts},
\link{md_legislative_districts}, or \link{md_senate_districts},
Default: NULL}

\item{type}{Type of data to return, Default: "counties"; See details for supported options.}

\item{...}{Additional parameters passed on to {tigris} functions.}
}
\value{
A simple feature object matching the type provided.
}
\description{
Get package data with Maryland counties, U.S. Congressional District, or Census designated
places by name or GeoID or use the tigris package to download state data.
}
\details{
Use {mapmaryland} packaged data or use tigris U.S. Census Bureau API:

Supported options that access data included with the {mapmaryland} package
include "counties", "census places", "congressional districts", "legislative
districts", and "water".

Options that access data by calling one of several tigris functions for
downloading data from the U.S. Census Bureau API include "senate district",
"tracts", "block groups", "blocks", "pumas", "voting districts". tigris
functions that do not use a "state" parameter (e.g. \link[tigris:zctas]{tigris::zctas})  are not
supported by this function.
}
