% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapmaryland-package.R
\docType{package}
\name{mapmaryland-package}
\alias{mapmaryland}
\alias{mapmaryland-package}
\title{mapmaryland: Easy Access to Maryland Spatial Data}
\description{
A small collection of data sources and utility functions for working with administrative and other data sources in Maryland.
}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID})

}
\keyword{internal}
