% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_arts_districts}
\alias{md_arts_districts}
\title{Maryland Arts & Entertainment Districts}
\format{
A data frame with 29 rows and 12 variables:
\describe{
\item{\code{name}}{District name}
\item{\code{desc}}{District description}
\item{\code{org_name}}{Organization name}
\item{\code{org_url}}{Organization website URL}
\item{\code{org_phone}}{Organization phone number}
\item{\code{org_address}}{Organization street address}
\item{\code{org_city}}{Organization city}
\item{\code{org_county}}{Organization county}
\item{\code{zipcode}}{Organization zipcode}
\item{\code{lon}}{Longitude}
\item{\code{lat}}{Latitude}
\item{\code{geometry}}{MULTIPOLYGON geometry with boundaries}
}
}
\source{
\url{https://geodata.md.gov/imap/rest/services/BusinessEconomy/MD_IncentiveZones/FeatureServer/2}
}
\usage{
md_arts_districts
}
\description{
The Maryland State Arts Council sponsors Arts & Entertainment Districts as a
way to stimulate the economy and improve quality of life. This legislation
enables local jurisdictions, municipalities, counties, or a combination
thereof, to apply for state designations for the Arts and Entertainment
Districts within their boundaries and offer tax incentives as provided by
law. More information from the Maryland State Arts Council:
\url{https://msac.org/programs/arts-entertainment-districts}
}
\keyword{datasets}
