% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_census_places}
\alias{md_census_places}
\title{Maryland Census-designated places}
\format{
A data frame with 536 rows and 17 variables:
\describe{
\item{\code{statefp}}{State FIPS code for Maryland}
\item{\code{county}}{County name}
\item{\code{countyfp}}{County FIPS code}
\item{\code{placefp}}{CDP FIPS code}
\item{\code{placens}}{CDP GNIS code}
\item{\code{geoid}}{Unique CDP GeoID}
\item{\code{name}}{Place name}
\item{\code{namelsad}}{Place Legal/statistical area description (LSAD)}
\item{\code{lsad}}{Legal/statistical area description (LSAD)}
\item{\code{classfp}}{Class FIPS code}
\item{\code{pcicbsa}}{Current metropolitan/micropolitan statistical area principal city indicator}
\item{\code{mtfcc}}{MAF/TIGER Feature Class Code (MTFCC)}
\item{\code{funcstat}}{Functional status}
\item{\code{aland}}{Land area (square meters)}
\item{\code{awater}}{Water area (square meters)}
\item{\code{intptlat}}{Latitude of the internal point}
\item{\code{intptlon}}{Longitude of the internal point}
\item{\code{geometry}}{list COLUMN_DESCRIPTION}
}
}
\usage{
md_census_places
}
\description{
Census Designated Places (CDPs) in the state of Maryland downloaded using the
\link[tigris:places]{tigris::places} function.
}
\keyword{datasets}
