% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_congressional_districts}
\alias{md_congressional_districts}
\title{Maryland U.S. Congressional Districts}
\format{
A data frame with 8 rows and 13 variables:
\describe{
\item{\code{statefp}}{2-character state FIPS code}
\item{\code{cd116fp}}{116th congressional district FIPS code}
\item{\code{geoid}}{GeoID}
\item{\code{namelsad}}{concatenated variable with geographic area name and legal/statistical area description (LSAD)}
\item{\code{lsad}}{Legal/statistical area description (LSAD)}
\item{\code{cdsessn}}{Congressional session code}
\item{\code{mtfcc}}{MAF/TIGER Feature Class Code (MTFCC)}
\item{\code{funcstat}}{functional status}
\item{\code{aland}}{land area (square meters)}
\item{\code{awater}}{water area (square meters)}
\item{\code{intptlat}}{latitude of the internal point}
\item{\code{intptlon}}{longitude of the internal point}
\item{\code{label}}{Congressional District label}
\item{\code{name}}{Congressional District name}
\item{\code{geometry}}{Multipolygon with district boundary}
}
}
\usage{
md_congressional_districts
}
\description{
U.S. Congressional Districts boundaries downloaded from the U.S. Census Bureau with the
\link[tigris:congressional_districts]{tigris::congressional_districts} function. Updated 2024-10-29.
}
\keyword{datasets}
