% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_counties}
\alias{md_counties}
\title{Maryland Counties (boundaries)}
\format{
A data frame with 24 rows and 18 variables:
\describe{
\item{\code{statefp}}{State FIPS code for Maryland}
\item{\code{countyfp}}{County FIPS code}
\item{\code{countyns}}{County GNIS code}
\item{\code{geoid}}{Unique county FIPS code (concatenation of state and county FIPS codes)}
\item{\code{name}}{County name}
\item{\code{namelsad}}{Concatenated variable length geographic area name and
legal/statistical area description (LSAD)}
\item{\code{lsad}}{Legal/statistical area description (LSAD)}
\item{\code{classfp}}{FIPS class code}
\item{\code{mtfcc}}{MAF/TIGER Feature Class Code (MTFCC)}
\item{\code{csafp}}{Combined statistical area code}
\item{\code{cbsafp}}{Metropolitan statistical area/micropolitan statistical area code}
\item{\code{metdivfp}}{Metropolitan division code}
\item{\code{funcstat}}{Functional status}
\item{\code{aland}}{Land area (square meters)}
\item{\code{awater}}{Water area (square meters)}
\item{\code{intptlat}}{Latitude of the internal point}
\item{\code{intptlon}}{Longitude of the internal point}
\item{\code{region}}{Region of the state}
\item{\code{geometry}}{Multipolygon with the county boundary}
}
}
\source{
\url{https://www.census.gov/geo/maps-data/data/tiger-line.html}
}
\usage{
md_counties
}
\description{
County boundaries downloaded from the U.S. Census Bureau with the
\link[tigris:counties]{tigris::counties} function. Counties in the Baltimore–Columbia–Towson
Metropolitan Statistical Area (MSA) include Baltimore City, Baltimore County,
Carroll County, Anne Arundel County, Howard County, Queen Anne's County, and
Harford County. Counties in the DC-VA-MD-WV Metropolitan Statistical Area
(MSA) include Montgomery, Frederick, Prince George's, Charles, and Calvert
counties. Regions are based on this resource from Visit Maryland
\url{https://www.visitmaryland.org/info/maryland-regions}
}
\keyword{datasets}
