% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_counties_detailed}
\alias{md_counties_detailed}
\title{Maryland Counties (detailed boundaries)}
\format{
A data frame with 24 rows and 6 variables:
\describe{
\item{\code{statefp}}{State FIPS code for Maryland}
\item{\code{countyfp}}{County FIPS code}
\item{\code{name}}{County name}
\item{\code{geoid}}{Unique county GeoID (concatenation of state and county FIPS codes)}
\item{\code{namelsad}}{Concatenated variable length geographic area name and legal/statistical area description (LSAD)}
\item{\code{geometry}}{Multipolygon with the county boundary}
}
}
\usage{
md_counties_detailed
}
\description{
Detailed boundaries for Maryland counties from Maryland iMap.
}
\keyword{datasets}
