% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_imap_index}
\alias{md_imap_index}
\title{Maryland iMap Folder/Service Index}
\format{
A data frame with 1,505 rows and 16 variables:
\describe{
\item{\code{name}}{character Name}
\item{\code{nm}}{character Name with snake case}
\item{\code{type}}{character Service/layer type}
\item{\code{url}}{character Folder/service/layer URL}
\item{\code{urlType}}{character Index type}
\item{\code{folderPath}}{Folder path}
\item{\code{serviceName}}{character Service/layer type}
\item{\code{serviceType}}{character Service/layer type}
\item{\code{id}}{integer Layer ID number}
\item{\code{parentLayerId}}{integer Parent layer ID number}
\item{\code{defaultVisibility}}{logical Layer default visibility}
\item{\code{minScale}}{double Minimum scale}
\item{\code{maxScale}}{integer Maximum scale}
\item{\code{geometryType}}{character Geometry type}
\item{\code{subLayerIds}}{list Sublayer ID numbers}
\item{\code{supportsDynamicLegends}}{logical Supports dynamic legends}
}
}
\usage{
md_imap_index
}
\description{
Created using esriIndex function (in process of being developed for esri2sf
package). Updated August 22, 2023.
}
\keyword{datasets}
