% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_legislative_districts}
\alias{md_legislative_districts}
\title{Maryland State Legislative Districts}
\format{
A data frame with 68 rows and 13 variables:
\describe{
\item{\code{statefp}}{2-character state FIPS code}
\item{\code{sldlst}}{State Legislative District Lower Chamber}
\item{\code{geoid}}{GeoID}
\item{\code{namelsad}}{concatenated variable with geographic area name and legal/statistical area description (LSAD)}
\item{\code{lsad}}{Legal/statistical area description (LSAD)}
\item{\code{lsy}}{Legislative session year}
\item{\code{mtfcc}}{MAF/TIGER Feature Class Code (MTFCC)}
\item{\code{funcstat}}{functional status}
\item{\code{aland}}{land area (square meters)}
\item{\code{awater}}{water area (square meters)}
\item{\code{intptlat}}{latitude of the internal point}
\item{\code{intptlon}}{longitude of the internal point}
\item{\code{geometry}}{Multipolygon with district boundary}
}
}
\usage{
md_legislative_districts
}
\description{
Maryland State Legislative Districts boundaries downloaded from the U.S. Census Bureau with the
\link[tigris:state_legislative_districts]{tigris::state_legislative_districts} function with \code{house = "lower"}.
}
\keyword{datasets}
