% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_mpos}
\alias{md_mpos}
\title{Maryland Metropolitan Planning Organizations (boundaries)}
\format{
A data frame with 7 rows and 6 variables:
\describe{
\item{\code{geoid}}{GeoID}
\item{\code{name}}{MPO Name}
\item{\code{abb}}{MPO abbreviation or acronym}
\item{\code{url}}{MPO website url}
\item{\code{states}}{States (multiple states separated by semi-colons)}
\item{\code{geometry}}{MULTIPOLYGON geometry for boundaries}
}
}
\usage{
md_mpos
}
\description{
Data from \url{https://geodata.md.gov/imap/rest/services/BusinessEconomy/MD_IncentiveZones/FeatureServer/14}
}
\keyword{datasets}
