% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_open_data_index}
\alias{md_open_data_index}
\title{Maryland Open Data Portal Index}
\format{
A data frame with 737 rows and 11 variables:
\describe{
\item{\code{name}}{Resource name}
\item{\code{nm}}{Resource name (snake case)}
\item{\code{resource}}{Resource identifier}
\item{\code{description}}{Description}
\item{\code{url}}{Resource URL}
\item{\code{issued}}{Date issued}
\item{\code{modified}}{Date modified}
\item{\code{keyword}}{list of keywords}
\item{\code{identifier}}{API identifier}
\item{\code{theme}}{Data theme}
\item{\code{license}}{Data license}
}
}
\usage{
md_open_data_index
}
\description{
Public resources from the Maryland Open Data Portal excluding resources
cross-listed from Maryland iMap.
}
\keyword{datasets}
