% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_schools}
\alias{md_schools}
\title{Maryland K-12 Schools (Public and Charter Schools)}
\format{
A data frame with 1,427 rows and 13 variables:
\describe{
\item{\code{school_name}}{School name}
\item{\code{address}}{Street address}
\item{\code{city}}{City name}
\item{\code{state}}{State}
\item{\code{zip}}{Zip code}
\item{\code{grades}}{Grades}
\item{\code{school_type}}{School type}
\item{\code{grade_band}}{Grade band}
\item{\code{psc_number}}{PSC number}
\item{\code{lss_number}}{Local school system number as integer}
\item{\code{county}}{County name}
\item{\code{management_type}}{Management type}
\item{\code{geometry}}{POINT geometry}
}
}
\usage{
md_schools
}
\description{
Combined iMap FeatureLayer for traditional public schools and charter
schools. The school number previously included in this dataset was incomplete
due to the iMap data using a different school name than the MSDE data for a
little less than half of schools in the state. As of February 2023, the
school number column has been removed to ensure the data is accurate and to
avoid confusion. For charter schools, the school_type column has been added
based on the grade range. Data based on iMap data layers as of 2023 February
10 for public schools
\url{https://geodata.md.gov/imap/rest/services/Education/MD_EducationFacilities/FeatureServer/5}
and charter schools
\url{https://geodata.md.gov/imap/rest/services/Education/MD_EducationFacilities/FeatureServer/6}
}
\keyword{datasets}
