% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_senate_districts_2022}
\alias{md_senate_districts_2022}
\title{Maryland Senate Districts (2022)}
\format{
A data frame with 47 rows and 5 variables:
\describe{
\item{\code{name}}{District name}
\item{\code{id}}{District number}
\item{\code{label}}{District label}
\item{\code{pop}}{Population from 2020 U.S. Census}
\item{\code{geometry}}{District boundary geometry}
}
}
\source{
\href{https://planning.maryland.gov/Redistricting/Pages/2020/legiDist.aspx}{2022 Maryland Legislative Districts}
}
\usage{
md_senate_districts_2022
}
\description{
From the \href{https://planning.maryland.gov/Redistricting/Pages/2020/legiDist.aspx}{Maryland Department of Planning}: "The Maryland General Assembly,
under the Maryland Constitution, adopted new state legislative districts on
February 1, 2022, based on the changes in population reported in the 2020
U.S. Census and adjusted in accordance with Maryland's \emph{No Representation
Without Population Act} of 2010."
}
\keyword{datasets}
