% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_streams}
\alias{md_streams}
\title{Maryland Rivers and Streams}
\format{
A data frame with 347 rows and 12 variables:
\describe{
\item{\code{objectid}}{integer GIS Object ID}
\item{\code{huc}}{character Hydrologic Unit}
\item{\code{lev}}{double Stream Level}
\item{\code{pname}}{character Name}
\item{\code{owname}}{character Open Water Name}
\item{\code{edaname}}{character Estuarine Coastal Drainage Area Name}
\item{\code{huc2}}{double HUC 2 Digit}
\item{\code{huc4}}{double HUC 4 Digit}
\item{\code{huc6}}{double HUC 6 Digit}
\item{\code{countyfp}}{County FIPS}
\item{\code{county}}{County name}
\item{\code{geometry}}{Multipolygon geometry}
}
}
\source{
\url{https://geodata.md.gov/imap/rest/services/Hydrology/MD_Waterbodies/FeatureServer/0}
}
\usage{
md_streams
}
\description{
Maryland Rivers and Streams
}
\keyword{datasets}
