% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_streams_detailed}
\alias{md_streams_detailed}
\title{Maryland Rivers and Streams (detailed)}
\format{
A data frame with 95507 rows and 4 variables:
\describe{
\item{\code{type}}{character Type}
\item{\code{countyfp}}{character County FIPS}
\item{\code{county}}{character County name}
\item{\code{geometry}}{list Multipolygon geometry}
}
}
\source{
\url{https://geodata.md.gov/imap/rest/services/Hydrology/MD_Waterbodies/FeatureServer/2}
}
\usage{
md_streams_detailed
}
\description{
Maryland Rivers and Streams (detailed)
}
\keyword{datasets}
