% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_water}
\alias{md_water}
\title{Maryland Water (area)}
\format{
A data frame with 14,970 rows and 9 variables:
\describe{
\item{\code{ansicode}}{character ANSI code}
\item{\code{hydroid}}{character Hydroid}
\item{\code{fullname}}{character Full name}
\item{\code{mtfcc}}{character MAF/TIGER Feature Class Code}
\item{\code{aland}}{double Land area}
\item{\code{awater}}{double Water area}
\item{\code{intptlat}}{character Interior latitude point}
\item{\code{intptlon}}{character Interior longitude point}
\item{\code{geometry}}{list Multipolygon geometry}
}
}
\source{
\url{https://data.imap.maryland.gov/datasets/maryland-waterbodies-rivers-and-streams-detailed}
}
\usage{
md_water
}
\description{
Detailed multipolygon data for streams, lakes, and other water in the state
of Maryland downloaded from the U.S. Census Bureau with the
\link[tigris:area_water]{tigris::area_water} function.
}
\keyword{datasets}
