% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_zctas}
\alias{md_zctas}
\title{Maryland Zip Code Tabulation Areas (ZCTA) - 2020}
\format{
A data frame with 478 rows and 10 variables:
\describe{
\item{\code{zcta5ce20}}{ZCTA identifier}
\item{\code{geoid20}}{ZCTA GeoID}
\item{\code{classfp20}}{FIPS class code}
\item{\code{mtfcc20}}{MAF/TIGER Feature Class Code (MTFCC)}
\item{\code{funcstat20}}{Functional status}
\item{\code{aland20}}{Land area (square meters)}
\item{\code{awater20}}{Water area (square meters)}
\item{\code{intptlat20}}{Latitude of the internal point}
\item{\code{intptlon20}}{Longitude of the internal point}
\item{\code{geometry}}{Multipolygon with the ZCTA boundary}
}
}
\source{
\href{https://www.census.gov/programs-surveys/geography/guidance/geo-areas/zctas.html}
}
\usage{
md_zctas
}
\description{
ZIP Code Tabulation Areas (ZCTAs) are generalized areal representations of
United States Postal Service (USPS) ZIP Code service areas downloaded using
the \code{\link[tigris:zctas]{tigris::zctas()}} function.
}
\keyword{datasets}
