% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{real_property_cols}
\alias{real_property_cols}
\title{Maryland Real Property Assessments: Fields Reference}
\format{
A data frame with 223 rows and 6 variables:
\describe{
\item{\code{field_order}}{Field order}
\item{\code{field_name}}{Field name}
\item{\code{data_type}}{Data type}
\item{\code{mdp_field_name}}{Maryland Department of Planning (MDP) Field name}
\item{\code{api_field_name}}{Maryland Open Data portal API field name}
\item{\code{sdat_field_number}}{State Department of Assessments and Taxation (SDAT) field number}
}
}
\source{
\url{https://opendata.maryland.gov/Business-and-Economy/Maryland-Real-Property-Assessments-Fields-Referenc/w8th-47fz/data}
}
\usage{
real_property_cols
}
\description{
Maryland Real Property Assessments: Fields Reference
}
\keyword{datasets}
