% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_mapmaryland_option}
\alias{set_mapmaryland_option}
\title{Set options for mapmaryland package}
\usage{
set_mapmaryland_option(..., overwrite = TRUE)
}
\arguments{
\item{...}{options to set, e.g. "crs = 2804" to set "mapmaryland.crs" to 2804.}

\item{overwrite}{If \code{TRUE}, overwrite any existing option value.}
}
\description{
Currently only used for mapmaryland.crs
}
