% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_md_data.R
\name{format_md_sf}
\alias{format_md_sf}
\title{Format Maryland simple feature data}
\usage{
format_md_sf(
  data,
  crs = 3857,
  erase_water = FALSE,
  clean_names = TRUE,
  sf_col = "geometry"
)
}
\arguments{
\item{data}{A \code{sf} object.}

\item{crs}{Coordinate reference system passed to \link[sf:st_transform]{sf::st_transform}, Default: 3857}

\item{erase_water}{If \code{TRUE}, use \link[overedge:st_erase]{overedge::st_erase} to erase any geometry intersecting with \link{md_water}; Default: \code{FALSE}}

\item{clean_names}{If \code{TRUE}, use \link[janitor:clean_names]{janitor::clean_names}; Default: \code{TRUE}}

\item{sf_col}{Name to use for the sf column after renaming; defaults to "geometry".}
}
\value{
A modified version of the input simple feature data.
}
\description{
Convert the coordinate reference system, convert invalid geometry to valid
geometry with \link[sf:valid]{sf::st_make_valid}, and, optionally, erase water, and clean
names with \link[janitor:clean_names]{janitor::clean_names}.
}
\seealso{
\link[overedge:format_data]{overedge::format_data}
}
