% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_md_crash_data.R
\name{get_md_crash_data}
\alias{get_md_crash_data}
\alias{format_md_crash_data}
\alias{format_md_crashes}
\alias{format_md_crash_date}
\alias{format_md_crashes_person}
\title{Get and format Maryland crash data}
\usage{
get_md_crash_data(
  location = NULL,
  ...,
  report_no = NULL,
  where = NULL,
  type = "crashes"
)

format_md_crash_data(data, type = "crashes", drop_code = TRUE, ...)

format_md_crashes(data)

format_md_crash_date(data, cols = c("acc_date", "acc_time"))

format_md_crashes_person(
  data,
  start_year = 2022,
  end_year = 2022,
  na_dates = c("1/1/1900", "19000101", "19001111", "19001212", "19200202")
)
}
\arguments{
\item{...}{Passed to \code{\link[=get_md_open_data]{get_md_open_data()}} for \code{\link[=get_md_crash_data]{get_md_crash_data()}} or passed
to specified formatting functions when used with \code{\link[=format_md_crash_data]{format_md_crash_data()}}.}

\item{report_no}{Crash report numbers. Defaults to \code{NULL}. Optionally used to
filter results if type is "persons" or "vehicles".}

\item{type}{Supported options include "crashes" (default), "persons" (or
"crashes_person"), and "vehicles" (or "crashes_vehicle").}

\item{data}{Data frame with Maryland vehicular crash data; typically from \code{get_md_crash_data}}

\item{drop_code}{If \code{TRUE} (default), drop all columns that end with "code"}

\item{cols}{Column names to use with \code{\link[=format_md_crash_date]{format_md_crash_date()}}.}
}
\description{
Get crash data using \code{\link[=get_md_open_data]{get_md_open_data()}}. \link{format_md_crash_data} works with
a data frame of crash data from \code{\link[=get_md_crash_data]{get_md_crash_data()}} when type = "crashes".
\code{\link[=format_md_crashes_person]{format_md_crashes_person()}} formats data when type = "persons".
\code{\link[=format_md_crashes]{format_md_crashes()}} calls one or the other depending on the type
parameter.
}
\examples{
\dontrun{
if (interactive()) {
  get_md_crash_data(
    where = "(year = '2020') AND (quarter = 'Q2')",
    name_col = "county_desc",
    name = "Cecil"
  )
}
}
}
