% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_data.R
\name{get_md_open_data}
\alias{get_md_open_data}
\alias{get_md_crash_data}
\title{Get data from the Maryland Open Data Portal}
\usage{
get_md_open_data(
  resource = NULL,
  type = NULL,
  crs = getOption("mapmaryland.crs", default = 3857),
  geometry = FALSE,
  token = NULL,
  ...
)

get_md_crash_data(...)
}
\arguments{
\item{resource}{Resource identifier code passed to "data" parameter of
\link[getdata:get_open_data]{getdata::get_open_data}}

\item{type}{Type of data to return. Supported options: "crashes", "road
closures", "bay pollution reduction"}

\item{crs}{Coordinate reference system of bounding box to return}

\item{geometry}{If \code{TRUE} and coords are provided, return a
\code{sf} object. Default \code{FALSE}.}

\item{...}{
  Arguments passed on to \code{\link[getdata:get_open_data]{getdata::get_open_data}}
  \describe{
    \item{\code{data}}{A data set identifier (known as a resource for Socrata) or a url
for an individual dataset. If data is set to "list" and a valid source_url
is provided, the function returns a list of all available resources. If
data is a url, source_url must be NULL. \link[getdata]{get_socrata_metadata} requires the
data parameter.}
    \item{\code{source_url}}{A data source url. For Socrata, this should the base url
for the open data portal.}
    \item{\code{source_type}}{Data source type; defaults to "socrata" which is currently
the only supported option.}
    \item{\code{select}}{SODA $select parameter. Set of columns to be returned, equivalent
to a SELECT in SQL. \url{https://dev.socrata.com/docs/queries/select.html}}
    \item{\code{where}}{SODA $where parameter. Filters the rows to be returned, equivalent
to WHERE in SQL. \url{https://dev.socrata.com/docs/queries/where.html}}
    \item{\code{query}}{SODA $query parameter. A full SoQL query string, all as one
parameter. \url{https://dev.socrata.com/docs/queries/query.html}}
    \item{\code{location}}{sf object. If multiple areas are provided, they are unioned
into a single sf object using \link[sf:geos_combine]{sf::st_union}}
    \item{\code{dist}}{buffer distance in units. Optional.}
    \item{\code{diag_ratio}}{ratio of diagonal distance of area's bounding box used as
buffer distance. e.g. if the diagonal distance is 3000 meters and the
"diag_ratio = 0.1"  a 300 meter will be used. Ignored when \code{dist} is
provided.}
    \item{\code{unit}}{Units for buffer. Supported options include "meter", "foot",
"kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km"
instead of "kilometer") are also supported. Distance in units is converted
to units matching GDAL units for x; defaults to "meter"}
    \item{\code{asp}}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
character (e.g. "1:3"). If numeric, \code{\link[sfext:get_asp]{get_asp()}} returns the same value
without modification.}
    \item{\code{name,name_col}}{Name of column in Socrata data resource with
location names (e.g. County) and name of location to return.}
    \item{\code{location_col}}{Name of a "location" or "point" type column in a Socrata dataset.}
    \item{\code{coords}}{Coordinate columns for input dataframe or output sf object (if
geometry is 'centroid' or 'point') Default: c("lon", "lat").}
    \item{\code{token,type}}{Access token or API Key and token type (name used to store
token in .Renvironment). A token may be required to access data from
Socrata and other open data portals but can be stored as an environment
variable with \link[getdata]{set_access_token}.}
    \item{\code{from_crs}}{Coordinate reference system used to match the location CRS to
the source data.}
    \item{\code{clean_names}}{If \code{TRUE}, clean names provided to nm or created based on
value of col using \link[janitor:clean_names]{janitor::clean_names}. If \code{FALSE}, use names as
provided.}
  }}
}
\description{
Wrap \link[getdata:get_open_data]{getdata::get_open_data} to access the Maryland Open Data portal. The
source url can be set with an option named "mapmaryland.open_data_source_url"
but it defaults to the statewide open data portal
\url{https://opendata.maryland.gov} The token type can be set with an option
named "mapmaryland.open_data_token_type" but it defaults to
"MARYLAND_OPEN_DATA_API_KEY".
}
\examples{
\dontrun{
if (interactive()) {
  get_md_crash_data(
    where = "(year = '2020') AND (quarter = 'Q2')",
    name_col = "county_desc",
    name = "Cecil"
  )
}
}
}
