% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_data.R
\name{get_md_open_data}
\alias{get_md_open_data}
\alias{get_md_crash_data}
\title{Get data from the Maryland Open Data Portal}
\usage{
get_md_open_data(
  resource = NULL,
  type = NULL,
  crs = getOption("mapmaryland.crs", default = 3857),
  geometry = FALSE,
  token = NULL,
  ...
)

get_md_crash_data(...)
}
\arguments{
\item{type}{Type of data to return. Supported options: "crashes", "road closures", "bay pollution reduction"}

\item{crs}{Coordinate reference system to return.}

\item{geometry}{If \code{TRUE} and latitude/longitude columns available, return a
\code{\link[=sf]{sf()}} object. Default \code{FALSE}.}

\item{token}{Access token or API Key; required to access data from Socrata.}
}
\description{
Get data from the Maryland Open Data Portal
}
\examples{
\dontrun{
if (interactive()) {
  get_md_crash_data(
    where = "(year = '2020') AND (quarter = 'Q2')",
    name_col = "county_desc",
    name = "Cecil"
  )
}
}
}
