% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapmaryland-package.R
\docType{package}
\name{mapmaryland-package}
\alias{mapmaryland}
\alias{mapmaryland-package}
\title{mapmaryland: Easy Access to Maryland Spatial Data}
\description{
mapmaryland helps you get open spatial data for Maryland.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/mapmaryland}
  \item \url{https://elipousson.github.io/mapmaryland/}
  \item Report bugs at \url{https://github.com/elipousson/mapmaryland/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
