% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{md_schools}
\alias{md_schools}
\title{Maryland Public Schools}
\format{
A data frame with 1,427 rows and 14 variables:
\describe{
\item{\code{school_name}}{School name}
\item{\code{address}}{Street address}
\item{\code{city}}{City name}
\item{\code{state}}{State}
\item{\code{zip}}{Zip code}
\item{\code{grades}}{Grades}
\item{\code{school_type}}{School type}
\item{\code{grade_band}}{Grade band}
\item{\code{lss_number}}{Local school system number}
\item{\code{psc_number}}{PSC number}
\item{\code{county}}{County name}
\item{\code{management_type}}{Management type}
\item{\code{school_number}}{School number (incomplete)}
\item{\code{geometry}}{POINT geometry}
}
}
\usage{
md_schools
}
\description{
Combined iMap FeatureLayer for traditional public schools and charter
schools. The school number very incomplete due to the iMap data using a
different school name than the MSDE data for a little less than half of
schools in the state.
}
\keyword{datasets}
